fix: replace unused humanizeEnumLabel with proper formatDisplay label helper
