+++ apps/sdp-web/src/app/dashboard/issuance/[tokenId]/token-overview-section.tsx
@@ -3,7 +3,8 @@
 import type { Token } from "@sdp/types";
 import { Button } from "@/components/ui/button";
 import { TokenDisabledActionTooltip } from "./token-disabled-action-tooltip";
-import { formatDate, formatDisplayLabel } from "./token-management-workspace.utils";
+import { formatDate } from "./token-management-workspace.utils";
+import { formatDisplayLabel } from "@/lib/utils";
 
 interface TokenOverviewSectionProps {
   token: Token;
