+++ apps/sdp-web/src/lib/utils.ts
@@ -5,8 +5,14 @@ export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
 
-export function humanizeEnumLabel(value: string): string {
-  const words = value.toLowerCase().split("_").filter(Boolean);
-  if (words.length === 0) return value;
-  return words.map((w, i) => (i === 0 ? w.charAt(0).toUpperCase() + w.slice(1) : w)).join(" ");
+const DISPLAY_LABEL_OVERRIDES: Record<string, string> = {
+  rwa: "RWA",
+};
+
+export function formatDisplayLabel(value: string): string {
+  const lower = value.toLowerCase();
+  if (DISPLAY_LABEL_OVERRIDES[lower]) return DISPLAY_LABEL_OVERRIDES[lower];
+
+  // tokenized-security => Tokenized Security, force_burn => Force Burn, rwa => RWA
+  return value.replace(/[_-]/g, " ").replace(/\b\w/g, (c) => c.toUpperCase());
 }
